[BUGFIX] Fix fetching more than 1000 existing resources on publish
When publishing a collection the existing resources in the target bucket
are fetched for later clean up. AWS SDK returns up to 1000 objects in one
request and offers pagination to fetch more. Using pagination parameter
with lowercase spelling causes S3 to always return the first page, providing
a marker for the next page, causing an endless loop. This fixes the parameter.